Moved RGBLedWrite to new file esp32-hal-rgb-led and created pinMode in variatn.cpp
