+++ cores/esp32/esp32-hal-gpio.h
@@ -72,9 +72,6 @@ extern "C" {
 #define digitalPinToDacChannel(pin)     (((pin) == DAC_CHANNEL_1_GPIO_NUM)?0:((pin) == DAC_CHANNEL_2_GPIO_NUM)?1:-1)
 
 void pinMode(uint8_t pin, uint8_t mode);
-#ifdef BOARD_HAS_NEOPIXEL
-  void RGBLedWrite(uint8_t pin, uint8_t red_val, uint8_t green_val, uint8_t blue_val);
-#endif
 void digitalWrite(uint8_t pin, uint8_t val);
 int digitalRead(uint8_t pin);
