+++ cores/esp32/esp32-hal.h
@@ -88,6 +88,7 @@ void yield(void);
 #include "esp32-hal-timer.h"
 #include "esp32-hal-bt.h"
 #include "esp32-hal-psram.h"
+#include "esp32-hal-rgb-led.h"
 #include "esp32-hal-cpu.h"
 
 void analogWrite(uint8_t pin, int value);
